Weatherforecast.py
Title: Weather Forecast Project in Python

Description:

The Weather Forecast Project in Python is a versatile and informative application designed to provide users with accurate and up-to-date weather information for a specified location. Leveraging external APIs and Python's capabilities, this project offers a user-friendly interface for accessing current weather conditions, forecasts, and other meteorological data.